Add simple input WMEs to Soar agent
I added the ability to initialize and update input-link WMEs for the Cozmo agent
from the CozmoSoar object. The WMEs currently working are the elements specified
in the README which do NOT have sub-bullets. In other words, the WMEs which do
not need an IdWME to create. The CozmoSoar.init_in_link method will create the
WMEs in question with the current values from the robot, and the
CozmoSoar.update_input method will update those WMEs with the current robot
values.

Both commit the changes but do NOT run another cycle, which is necessary for the
agent to see the changes. As such, when either method is called, it should be
followed by an Agent.RunSelf call.