Add turn-in-place action and refactor test Soar productions
In addition to adding the turn-in-place action, I changed up the test Soar
productions so that each action has its own sub-operator file and requires a
specific state ^name attribute to fire. This means that the test production
for each action has its own "environment" in some sense, and they don't
collide.